refactor: 파일 분리 및 함수 리팩토링
