+++ src/components/PaymentWidget.styles.ts
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+export const CardPaymentWidgetWrapper = styled.div`
+  background-color: white;
+  width: 376px;
+  height: 750px;
+  border-radius: 10px;
+`;
