+++ src/components/cardInfo/cardInfoHeader/CardInfoHeader.styles.ts
@@ -1,24 +1,8 @@
 import styled from "@emotion/styled";
 
-interface CardInfoHeaderProps {
-  guide: string;
-  subGuide: string;
-}
-
-export default function CardInfoHeader({
-  guide,
-  subGuide,
-}: CardInfoHeaderProps) {
-  return (
-    <CardInfoHeaderWrapper>
-      <h1>{guide}</h1>
-      <p>{subGuide}</p>
-    </CardInfoHeaderWrapper>
-  );
-}
-
-const CardInfoHeaderWrapper = styled.div`
+export const CardInfoHeaderWrapper = styled.div`
   gap: 4px;
+
   h1 {
     font-size: 18px;
     font-weight: 700;
@@ -27,6 +11,7 @@ const CardInfoHeaderWrapper = styled.div`
     vertical-align: middle;
     color: rgba(0, 0, 0, 1);
   }
+
   p {
     font-size: 9.5px;
     font-weight: 400;
