+++ src/components/cardInfo/cardInfoHeader/CardInfoHeader.tsx
@@ -1,0 +1,18 @@
+import { CardInfoHeaderWrapper } from "./CardInfoHeader.styles";
+
+interface CardInfoHeaderProps {
+  guide: string;
+  subGuide: string;
+}
+
+export default function CardInfoHeader({
+  guide,
+  subGuide,
+}: CardInfoHeaderProps) {
+  return (
+    <CardInfoHeaderWrapper>
+      <h1>{guide}</h1>
+      <p>{subGuide}</p>
+    </CardInfoHeaderWrapper>
+  );
+}
