+++ src/components/cardInfo/cardInfoInput/CardInfoInput.styles.ts
@@ -1,0 +1,24 @@
+import styled from "@emotion/styled";
+
+export const CardInfoInputWrapper = styled.div`
+  display: flex;
+  gap: 8px;
+  flex-direction: column;
+`;
+
+export const CardInfoInputLabel = styled.label`
+  font-size: 12px;
+  font-weight: 500;
+  letter-spacing: 0%;
+  line-height: 15px;
+  vertical-align: middle;
+  color: rgba(10, 13, 19, 1);
+`;
+
+export const InputContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+  width: 100%;
+`;
