+++ src/components/cardPreview/expireDate/PreviewCardExpireDate.styles.ts
@@ -1,0 +1,17 @@
+import styled from "@emotion/styled";
+
+export const CardNumberWrapper = styled.div`
+  display: flex;
+  font-size: 14px;
+  font-weight: 500;
+  letter-spacing: 16%;
+  line-height: 20px;
+  vertical-align: middle;
+  color: rgba(255, 255, 255, 1);
+  gap: "0px";
+`;
+
+export const CardNumberParagraph = styled.p`
+  margin: 0;
+  padding: 0;
+`;
