+++ src/components/cardPreview/expireDate/PreviewCardExpireDate.tsx
@@ -1,0 +1,17 @@
+import {
+  CardNumberWrapper,
+  CardNumberParagraph,
+} from "./PreviewCardExpireDate.styles";
+
+export default function PreviewCardExpireDate({
+  expireDate,
+}: {
+  expireDate: string[];
+}) {
+  const formatted = expireDate.filter(Boolean).join("/");
+  return (
+    <CardNumberWrapper>
+      <CardNumberParagraph>{formatted}</CardNumberParagraph>
+    </CardNumberWrapper>
+  );
+}
